src/main/java/com/leetcode/solution/problem1117/H2O.java
package com.leetcode.solution.problem1117;

import java.util.concurrent.Semaphore;

/**
 * H2O 生成
 *
 * @author Ricky
 * @date 2020-06-19 20:17:35
 */
public class H2O {


    private Semaphore h = new Semaphore(2);
    private Semaphore o = new Semaphore(0);

    public H2O() {
    }

    public void hydrogen(Runnable releaseHydrogen) throws InterruptedException {
        h.acquire();
        // releaseHydrogen.run() outputs "H". Do not change or remove this line.
        releaseHydrogen.run();
        o.release();
    }

    public void oxygen(Runnable releaseOxygen) throws InterruptedException {
        o.acquire(2);
        // releaseOxygen.run() outputs "O". Do not change or remove this line.
        releaseOxygen.run();
        h.release(2);
    }
}
